requirements/base.txt
alembic
celery[redis]
fastapi
httpx
uvicorn
pydantic
pyjwt
sqlalchemy